Add CI for osx, linux and windows

+++ src/posix/operations_posix.cpp
@@ -425,6 +425,30 @@ namespace {
 void
 remove_dir_rec(const path& p, std::uintmax_t& count, std::error_code& ec)
 {
+  const auto check_dirent_is_dir = [](
+    const path& subp, struct dirent* dp, std::error_code& ec1) {
+    if (dp->d_type == DT_UNKNOWN) {
+      // the filesystem doesn't support reporting the filetype via
+      // dirent.  Use ::lstat() as a fallback.
+      struct stat buf;
+      if (::lstat(subp.c_str(), &buf) < 0) {
+        ec1 = std::error_code(errno, std::generic_category());
+        return false;
+      }
+
+      ec1.clear();
+      return S_ISDIR(buf.st_mode);
+    }
+    else if (dp->d_type == DT_DIR) {
+      ec1.clear();
+      return true;
+    }
+    else {
+      ec1.clear();
+      return false;
+    }
+  };
+
   DIR* dirp = ::opendir(p.c_str());
   if (dirp) {
     struct dirent* direntp = nullptr;
@@ -443,13 +467,16 @@ remove_dir_rec(const path& p, std::uintmax_t& count, std::error_code& ec)
         }
 
         const auto subpath = p / direntp->d_name;
-        if (direntp->d_type == DT_DIR) {
+        if (check_dirent_is_dir(subpath, direntp, ec)) {
           remove_dir_rec(subpath, count, ec);
           if (ec) {
             return;
           }
         }
         else {
+          if (ec) {
+            return;
+          }
           if (::unlink(subpath.c_str())) {
             ec = std::error_code(errno, std::generic_category());
             return;
@@ -459,7 +486,10 @@ remove_dir_rec(const path& p, std::uintmax_t& count, std::error_code& ec)
       }
     } while (direntp);
 
-    ::closedir(dirp);
+    if (::closedir(dirp) < 0) {
+      ec = std::error_code(errno, std::generic_category());
+      return;
+    }
 
     if (::rmdir(p.c_str())) {
       ec = std::error_code(errno, std::generic_category());
@@ -481,9 +511,17 @@ remove_all(const path& p, std::error_code& ec) NOEXCEPT
 {
   std::uintmax_t count = 0;
 
-  remove_dir_rec(p, count, ec);
+  std::error_code ecp;
+  remove_dir_rec(p, count, ecp);
 
-  return ec ? static_cast<std::uintmax_t>(-1) : count;
+  if (ecp) {
+    ec = ecp;
+    return static_cast<std::uintmax_t>(-1);
+  }
+  else {
+    ec.clear();
+    return count;
+  }
 }
